// src/app/models/hackathon.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface Hackathon {
  id: number;
  title: string;
  description?: string;
  maxTeamSize: number;
  location?: string;
  isOnline?: boolean;
  logo?: string;
  teams?: Team[];
  createdBy?: User;
  startDate?: string;
  endDate?: string;
}

interface Team {
  id: number;
  teamName: string;
  teamMembers: User[];
}


interface User {
  id: number;
  name: string;
  lastname: string;
}